+++ server.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request, redirect
-from steamintegration import Steam
+from steam import Steam
 app = Flask(__name__)
 
 api_key = None
